Added Action Filter for MathController

+++ Service Layer/IMathService.cs
@@ -5,6 +5,6 @@
         int add(int x, int y);
         int subtract(int x, int y);
         int multiply(int x, int y);
-        double divide(double x, double y);
+        int divide(int x, int y);
     }
 }
