+++ Service Layer/MathService.cs
@@ -14,7 +14,7 @@
         {
             return x * y;
         }
-        public double divide(double x, double y)
+        public int divide(int x, int y)
         {
             return x / y;
         }
